+++ cupyx/scipy/fftpack/__init__.py
@@ -1,9 +1,9 @@
-from cupyx.scipy.fftpack.fft import fft  # NOQA
-from cupyx.scipy.fftpack.fft import fft2  # NOQA
-from cupyx.scipy.fftpack.fft import fftn  # NOQA
-from cupyx.scipy.fftpack.fft import ifft  # NOQA
-from cupyx.scipy.fftpack.fft import ifft2  # NOQA
-from cupyx.scipy.fftpack.fft import ifftn  # NOQA
-from cupyx.scipy.fftpack.fft import irfft  # NOQA
-from cupyx.scipy.fftpack.fft import rfft  # NOQA
-from cupyx.scipy.fftpack.fft import get_fft_plan  # NOQA
+from cupyx.scipy.fftpack._fft import fft  # NOQA
+from cupyx.scipy.fftpack._fft import fft2  # NOQA
+from cupyx.scipy.fftpack._fft import fftn  # NOQA
+from cupyx.scipy.fftpack._fft import ifft  # NOQA
+from cupyx.scipy.fftpack._fft import ifft2  # NOQA
+from cupyx.scipy.fftpack._fft import ifftn  # NOQA
+from cupyx.scipy.fftpack._fft import irfft  # NOQA
+from cupyx.scipy.fftpack._fft import rfft  # NOQA
+from cupyx.scipy.fftpack._fft import get_fft_plan  # NOQA
